Fixed handler registration behavior
Now two handlers registered to the same pattern will be treated as different
routes registrations(instead of being appended to the same handler array) if
the method used to register them was different.

On most cases this won't change
the current behavior, only when using 'all' mixed with some other registration
method.